refactor(ui): extract react flow nodes builder
